refactor(linear): extract issue activity helpers
