+++ src/main/core/linear/linear-issue-provider.ts
@@ -8,46 +8,12 @@ import { clampIssueLimit, normalizeSearchTerm } from '@main/core/issues/helpers/
 import type { IssueProvider } from '@main/core/issues/issue-provider';
 import { log } from '@main/lib/logger';
 import { linearConnectionService } from './linear-connection-service';
-
-type LinearCommentNode = {
-  id: string;
-  body: string;
-  createdAt: string;
-  updatedAt: string;
-  url: string;
-  user: { displayName: string; name: string } | null;
-};
-
-type LinearPageInfo = {
-  hasNextPage: boolean;
-  endCursor: string | null;
-};
-
-type LinearConnection<T> = {
-  nodes: T[];
-  pageInfo?: LinearPageInfo;
-};
-
-type LinearHistoryNode = {
-  id: string;
-  createdAt: string;
-  updatedAt: string;
-  actor: { displayName?: string | null; name?: string | null } | null;
-  fromState?: { name: string } | null;
-  toState?: { name: string } | null;
-  fromAssignee?: { displayName: string; name: string } | null;
-  toAssignee?: { displayName: string; name: string } | null;
-  fromProject?: { name: string } | null;
-  toProject?: { name: string } | null;
-  fromCycle?: { name: string } | null;
-  toCycle?: { name: string } | null;
-  fromPriority?: number | null;
-  toPriority?: number | null;
-  fromEstimate?: number | null;
-  toEstimate?: number | null;
-  fromTitle?: string | null;
-  toTitle?: string | null;
-};
+import {
+  formatLinearContext,
+  hydrateIssueActivity,
+  LINEAR_ISSUE_ACTIVITY_FIELDS,
+  type LinearIssueWithActivity,
+} from './linear-issue-activity';
 
 type LinearIssueSummaryNode = {
   id: string;
@@ -63,12 +29,7 @@ type LinearIssueSummaryNode = {
   updatedAt: string;
 };
 
-type LinearIssueNode = LinearIssueSummaryNode & {
-  comments: LinearConnection<LinearCommentNode>;
-  history: LinearConnection<LinearHistoryNode>;
-};
-
-const ACTIVITY_PAGE_SIZE = 50;
+type LinearIssueContextNode = LinearIssueWithActivity<LinearIssueSummaryNode>;
 
 const ISSUE_SUMMARY_FRAGMENT = `
   fragment IssueSummary on Issue {
@@ -83,54 +44,6 @@ const ISSUE_SUMMARY_FRAGMENT = `
     project { name }
     assignee { displayName name }
     updatedAt
-  }
-`;
-
-const ISSUE_COMMENTS_QUERY = `
-  query IssueComments($issueId: String!, $cursor: String) {
-    issue(id: $issueId) {
-      comments(first: ${ACTIVITY_PAGE_SIZE}, after: $cursor, orderBy: createdAt) {
-        pageInfo { hasNextPage endCursor }
-        nodes {
-          id
-          body
-          createdAt
-          updatedAt
-          url
-          user { displayName name }
-        }
-      }
-    }
-  }
-`;
-
-const ISSUE_HISTORY_QUERY = `
-  query IssueHistory($issueId: String!, $cursor: String) {
-    issue(id: $issueId) {
-      history(first: ${ACTIVITY_PAGE_SIZE}, after: $cursor, orderBy: createdAt) {
-        pageInfo { hasNextPage endCursor }
-        nodes {
-          id
-          createdAt
-          updatedAt
-          actor { ... on User { displayName name } }
-          fromState { name }
-          toState { name }
-          fromAssignee { displayName name }
-          toAssignee { displayName name }
-          fromProject { name }
-          toProject { name }
-          fromCycle { name }
-          toCycle { name }
-          fromPriority
-          toPriority
-          fromEstimate
-          toEstimate
-          fromTitle
-          toTitle
-        }
-      }
-    }
   }
 `;
 
@@ -185,195 +98,11 @@ const ISSUE_CONTEXT_QUERY = `
         project { name }
         assignee { displayName name }
         updatedAt
-        comments(first: ${ACTIVITY_PAGE_SIZE}, orderBy: createdAt) {
-          pageInfo { hasNextPage endCursor }
-          nodes {
-            id
-            body
-            createdAt
-            updatedAt
-            url
-            user { displayName name }
-          }
-        }
-        history(first: ${ACTIVITY_PAGE_SIZE}, orderBy: createdAt) {
-          pageInfo { hasNextPage endCursor }
-          nodes {
-            id
-            createdAt
-            updatedAt
-            actor { ... on User { displayName name } }
-            fromState { name }
-            toState { name }
-            fromAssignee { displayName name }
-            toAssignee { displayName name }
-            fromProject { name }
-            toProject { name }
-            fromCycle { name }
-            toCycle { name }
-            fromPriority
-            toPriority
-            fromEstimate
-            toEstimate
-            fromTitle
-            toTitle
-          }
-        }
+        ${LINEAR_ISSUE_ACTIVITY_FIELDS}
       }
     }
   }
 `;
-
-type NameLike = { displayName?: string | null; name?: string | null } | null | undefined;
-
-function displayName(user: NameLike, fallback: string): string;
-function displayName(user: NameLike): string | undefined;
-function displayName(user: NameLike, fallback?: string): string | undefined {
-  return user?.displayName ?? user?.name ?? fallback;
-}
-
-function formatTransition(
-  label: string,
-  from?: string | number | null,
-  to?: string | number | null
-) {
-  if (from === undefined && to === undefined) return undefined;
-  if (from === null && to === null) return undefined;
-  if (from === to) return undefined;
-  return `${label}: ${from ?? 'none'} -> ${to ?? 'none'}`;
-}
-
-function formatHistoryEntry(history: LinearHistoryNode): string {
-  const changes = [
-    formatTransition('State', history.fromState?.name, history.toState?.name),
-    formatTransition(
-      'Assignee',
-      displayName(history.fromAssignee),
-      displayName(history.toAssignee)
-    ),
-    formatTransition('Project', history.fromProject?.name, history.toProject?.name),
-    formatTransition('Cycle', history.fromCycle?.name, history.toCycle?.name),
-    formatTransition('Priority', history.fromPriority, history.toPriority),
-    formatTransition('Estimate', history.fromEstimate, history.toEstimate),
-    formatTransition('Title', history.fromTitle, history.toTitle),
-  ].filter(Boolean);
-
-  const summary = changes.length ? changes.join('; ') : 'Issue updated';
-  return `- ${history.createdAt} by ${displayName(history.actor, 'Unknown')}: ${summary}`;
-}
-
-function formatLinearContext(raw: LinearIssueNode): string | undefined {
-  const comments = raw.comments?.nodes ?? [];
-  const history = raw.history?.nodes ?? [];
-  if (comments.length === 0 && history.length === 0) return undefined;
-
-  const parts = ['Linear issue activity'];
-
-  if (comments.length > 0) {
-    parts.push(
-      '',
-      'Comments:',
-      ...comments.map(
-        (comment) =>
-          `- ${comment.createdAt} by ${displayName(comment.user, 'Unknown')}: ${comment.body.trim()}`
-      )
-    );
-  }
-
-  if (history.length > 0) {
-    parts.push('', 'History:', ...history.map(formatHistoryEntry));
-  }
-
-  return parts.join('\n');
-}
-
-type LinearRawClient = {
-  client: {
-    rawRequest: <TData, TVariables extends Record<string, unknown>>(
-      query: string,
-      variables: TVariables
-    ) => Promise<{ data?: TData }>;
-  };
-};
-
-function nextCursor(connection: LinearConnection<unknown> | undefined): string | undefined {
-  const pageInfo = connection?.pageInfo;
-  if (!pageInfo?.hasNextPage) return undefined;
-  return pageInfo.endCursor ?? undefined;
-}
-
-async function fetchRemainingComments(
-  client: LinearRawClient,
-  issueId: string,
-  cursor: string | undefined
-): Promise<LinearCommentNode[]> {
-  const comments: LinearCommentNode[] = [];
-  let pageCursor = cursor;
-
-  while (pageCursor) {
-    const { data } = await client.client.rawRequest<
-      { issue: { comments: LinearConnection<LinearCommentNode> } | null },
-      { issueId: string; cursor: string }
-    >(ISSUE_COMMENTS_QUERY, { issueId, cursor: pageCursor });
-
-    const page = data?.issue?.comments;
-    comments.push(...(page?.nodes ?? []));
-    pageCursor = nextCursor(page);
-  }
-
-  return comments;
-}
-
-async function fetchRemainingHistory(
-  client: LinearRawClient,
-  issueId: string,
-  cursor: string | undefined
-): Promise<LinearHistoryNode[]> {
-  const history: LinearHistoryNode[] = [];
-  let pageCursor = cursor;
-
-  while (pageCursor) {
-    const { data } = await client.client.rawRequest<
-      { issue: { history: LinearConnection<LinearHistoryNode> } | null },
-      { issueId: string; cursor: string }
-    >(ISSUE_HISTORY_QUERY, { issueId, cursor: pageCursor });
-
-    const page = data?.issue?.history;
-    history.push(...(page?.nodes ?? []));
-    pageCursor = nextCursor(page);
-  }
-
-  return history;
-}
-
-async function hydrateIssueActivity(
-  client: LinearRawClient,
-  issue: LinearIssueNode
-): Promise<LinearIssueNode> {
-  const commentsCursor = nextCursor(issue.comments);
-  const historyCursor = nextCursor(issue.history);
-
-  if (!commentsCursor && !historyCursor) return issue;
-
-  const [additionalComments, additionalHistory] = await Promise.all([
-    fetchRemainingComments(client, issue.id, commentsCursor),
-    fetchRemainingHistory(client, issue.id, historyCursor),
-  ]);
-
-  return {
-    ...issue,
-    comments: {
-      ...issue.comments,
-      nodes: [...(issue.comments?.nodes ?? []), ...additionalComments],
-      pageInfo: { hasNextPage: false, endCursor: null },
-    },
-    history: {
-      ...issue.history,
-      nodes: [...(issue.history?.nodes ?? []), ...additionalHistory],
-      pageInfo: { hasNextPage: false, endCursor: null },
-    },
-  };
-}
 
 function toIssue(raw: LinearIssueSummaryNode, context?: string): Issue {
   return {
@@ -464,7 +193,7 @@ async function getIssueContext(identifier: string): Promise<IssueContextResult> 
 
   try {
     const { data } = await client.client.rawRequest<
-      { searchIssues: { nodes: LinearIssueNode[] } },
+      { searchIssues: { nodes: LinearIssueContextNode[] } },
       { term: string; limit: number }
     >(ISSUE_CONTEXT_QUERY, {
       term,
